dbusgateway.cpp: handles multiple config calls
The dbus gateway used to put all of the config into its own config which
was later passed on to dbus-proxy. When called several times, like the
case if several capabilities that use the dbus gateway are requested,
this would fail with syntax error. The fix makes sure the array object
is not created until the config is sent to dbus proxy.