Fixed security risk setting env using snprintf
The config option "env-var-value" which promoted the syntax "prefix%ssuffix" is now removed. Instead there are two config options "env-var-prefix" and "env-var-suffix". Using this approach there is no need for snprintf hence the security threat is now removed.

For any system that has the "env-var-value" specified this must be changed.