container.h: use ReturnCode and pointer-paramenter
For most methods, we want to know whether or not the method succeeded,
and the ReturnCode type is ideal for this. It is now used everywhere
where we used to rely on having an optionally invalid pid returned.

The pid is passed as a pointer-parameter to the functions that set it.