removes dependency on pelagicore-utils
As part of the effort in making SoftwareContainer more OSS friendly,
this commit removes the dependency on pelagicore-utils, the cmake
macros and source code that comes with it. The cmake macros has been
copied into the actual cmake files where needed, and the source code
utils have been replaced or copy/pasted from the utils lib.

* add_gmock / add_gtest / add_coverage etc cmake macros are copied from utils.
* pelagicore-logging is replaced by using more or less "vanilla" ivi-logging.
* GlibDBusCppFactory template classes for adaptor/proxy have been copied
  into the classes where they are being used.
* CommandLineParser has been replaced by GNU getopts. It adds some boilerplate
  however in type checking.
* formatString has been replaced with simple string concatenation.