removes pulse dependencies, updates README
Since the pulse gateway has been rewritten to look for the PULSE_SERVER
variable instead of creating a socket by itself, there is neither a
compile-time nor runtime dependency on pulseaudio.

Also adds some reasoning behind the dependencies that we do have. There
was also a dead variable in the pc file for CMake.