Renamed SoftwareContainerLib and split Workspace
SoftwareContainerLib has been renamed to SoftwareContainer since the class itself is not a library and the naming is inconsistent.
The source files libsoftwarecontainer.cpp and .h are now called softwarecontainer.cpp and .h.

SoftwareContainerWorkspace has been renamed to Workspace and been split into a separate .h and .cpp file.
getDefaultWorkspace was not used by anything but one test in softwarecontainerlib_unittest.cpp and has been causing problems
and has therefore been removed.