Makes creation of network bridge an option
With this, SoftwareContainer will not create a network bridge at runtime
if one is missing. Instead, it will print an error message and exit. Creation
of network bridge is done through a compile-time option given to cmake.

A consequence of this check is that the rather ugly assert in the workspace
code is removed, so it is up to the creator of the workspace object to call
checkWorkspace. Furthermore, it is the responsibility of the creator of
the agent object to make it check for workspace (see main.cpp) before
triggering any preloading of containers.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>