The container ID set by Agent is used for LXC ID
Refactored how the ids for the LXC containers is set.
Now the agent sets the entire ID of the container instead of
only a prefix followed by a generated string.

However, if SoftwareContainer::init() is called before the
container ID is set an ID will ge generated prefixed by "SC-".