Make the service tests fail + store their results
The suspend test was not properly configured for read-only bind mount,
that is also fixed.

The run script for service tests now exit with exit status equal to
the sum of all exit statuses of the tests (so if one fail it will at
least be 1 or higher).

The Jenkinsfile is instructed to save the JUnit XML files from the service
tests.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>