Enables to run on D-Bus session bus.
The default is still to run on the system bus, but one can pass
an option on command-line to run on session bus instead.

Also, if the requested bus is either busy or cannot be registered
to, SoftwareContainer will fail instead of trying the other one,
which it did before in the case of the system bus.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>