Sets shutdown timeout to 1 sec, haltsignal to KILL
This sets the shutdown signal for containers to 1 second. At the same
time, that will not be applied directly, since this also sets the halt
signal to SIGKILL.

The background for this is that we have a bug where containers always
timeout when shutting down, leading to us calling stop() on the LXC
container, which in turn sends SIGKILL.

When the timeout bug is fixed, we should remove the haltsignal
directive completely, and then 1 second should still be plenty of time
to shut down.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>