Fixes spawn regression by only using one core
For some reason, when running with multiple cores and virtualbox,
performance degrades when stress testing by spawning many processes. An
article (http://www.mihaimatei.com/virtualbox-performance-issues-multiple-cpu-cores/)
suggests this is because processing with multicore in VBox requires the
physical cores it uses to be free (and locking more cpus take more
time).

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>