Netw-GW takes values from main config
Eventually, the config should probably be passed
to SoftwareContainer, but that has dependencies on some
other structural change which are not part of this commit.
As a middle step towards that, the agent now has the config
as a member but passes the network GW values directly
rather than sharing the config.

Signed-off-by: Joakim Gross <joakim.gross@pelagicore.com>