Vagrantfile: use icecc + bump cookbook + various
To use icecc we need to be available on the same network, hence
the network setup change.

Since we never use the VAGRANT_CPUS setting, I removed that too.

Also, instead of taking time to select the "best" mirror, I've
changed so that we only use the httpredir mirror, which more or
less does the same thing. So now we just do an apt-get update

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>