Gateway::setConfig now takes a json_t as input
It took std::string before, and through a long way of not updating
in all places when we update in one place, we ended up in a situation
where the library took a GatewayConfiguration object (which wraps a
bunch of json_t) and the SoftwareContainer class would convert this to a
string, which the Gateway class would convert back into json_t.

TL;DR: It is all json_t now.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>