refactor gateway creation, include container
Since we always have a container present when we create a gateway,
it makes little sense not to pass it as an argument to the gateway.

Also reworked the component tests for the library, since the previous
set up was quite bad. Now, gateways get a container, not a
softwarecontainer (which btw already contained all gateways), and there
is a nicer hierarchy in the test classes.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>