Fixes some logging in container + service test lib
The service test lib would pass an empty env var pair to the
agent as a default, instead of a truly empty dict.

The logging of environment variables sent to a function was
causing problems when being used with DLT, so it has been
reworked to print one at a time, and to not add any newlines
manually.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>