DeviceNodeGateway: Remove create device support
Remove support for creating new devices in the container via the
DeviceNodeGateway. Creating new devices is hard to fit into the
whitelisting strategy and creates more problems than it solves. As a
result of this change there is no longer any need specifying major
minor versions of the device in the config.

Signed-off-by: Viktor Sjölind <viktor.sjolind@pelagicore.com>